requirements.txt
astroid==2.4.2
autopep8==1.5.4
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
google-api-core==1.23.0
google-auth==1.23.0
google-cloud-language==2.0.0
googleapis-common-protos==1.52.0
grpcio==1.33.2
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
libcst==0.3.13
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.1.0
proto-plus==1.11.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pylint==2.6.0
PySocks==1.7.1
pytz==2020.4
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
toml==0.10.2
tweepy==3.9.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.25.11
wrapt==1.12.1
